+++ packages/json-comb-core/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.0.0] - 2018-03-13
+### Added
+- ✨ `opts.useNullAsExplicitFalse` on all both sync and async getKeyset methods.
+
+Since it's on by default (which is sensible), that's technically a breaking change, which warrants a major semver bump.
+
 ## [5.1.0] - 2018-01-27
 ### Added
 - ✨ `doNotFillThesePathsIfTheyContainPlaceholders` on both `enforceKeysetSync()` and `enforceKeyset`
@@ -97,3 +103,4 @@ API-wise, there are no changes, but since it's rewrite in ES modules and the arr
 [4.2.0]: https://github.com/codsen/json-comb-core/compare/v4.1.0...v4.2.0
 [5.0.0]: https://github.com/codsen/json-comb-core/compare/v4.2.0...v5.0.0
 [5.1.0]: https://github.com/codsen/json-comb-core/compare/v5.0.0...v5.1.0
+[6.0.0]: https://github.com/codsen/json-comb-core/compare/v5.1.3...v6.0.0